Script now counts the auto-allocated marks
BBL does not export UIDs for questions along with its .cvs output,
thereby making this hack necessary.

This problem arises specifically when students are given questions in a
random order - each column in the BBL output csv does not correspond to
a unique question.  No other identifiable information is provided
either.

This is not even funny anymore.